feat: 데이터 props로 컴포넌트간 연결
